Fix server exit if no conda found

+++ .config/zsh_custom/gypsum.zsh
@@ -1,20 +1,22 @@
-alias gps='ssh guanghaowei@gypsum.cs.umass.edu'
-alias gpsg='ssh guanghaowei@gypsum-gateway.cs.umass.edu'
+alias selenite='ssh guanghaowei@gypsum-gateway.cs.umass.edu'
+alias gypsum='ssh guanghaowei@gypsum.cs.umass.edu'
 
-alias gpcd='/mnt/nfs/work1/hongyu/guanghaowei'
+if [[ $(hostname) == node* ]]; then
+	alias gpcd='cd /mnt/nfs/work1/hongyu/guanghaowei'
 
-alias gpb='sbatch'
+	alias gpb='sbatch'
 
-alias gpsh='srun --pty /bin/zsh'
-alias gpshg='srun --pty --gres=gpu:1 /bin/zsh'
-alias gpshgl='srun --pty --gres=gpu:1 --partition=titanx-long /bin/zsh'
+	alias gpsh='srun --pty /bin/zsh'
+	alias gpshg='srun --pty --gres=gpu:1 /bin/zsh'
+	alias gpshgl='srun --pty --gres=gpu:1 --partition=titanx-long /bin/zsh'
 
-alias gpq='squeue | grep'
-alias gpps='squeue | grep guanghaowei'
+	alias gpq='squeue | grep'
+	alias gpps='squeue | grep guanghaowei'
 
-alias gpc='scancel'
+	alias gpc='scancel'
 
-alias gpi='sinfo'
+	alias gpi='sinfo'
+fi
 
 
 alias noteaid='cd ~/projects/NoteAid-Annotation; conda activate noteaid'
